backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-adminlogentries==0.1.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pygraphviz==1.5
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
